Problem: icaauth route is not registered correctly
test icaauth
