+++ integration_tests/cosmoscli.py
@@ -1276,6 +1276,28 @@ class CosmosCLI:
             rsp = self.event_query_tx_for(rsp["txhash"])
         return rsp
 
+    def icaauth_register_account(self, connid, **kwargs):
+        "execute on host chain to attach an account to the connection"
+        default_kwargs = {
+            "home": self.data_dir,
+            "node": self.node_rpc,
+            "chain_id": self.chain_id,
+            "keyring_backend": "test",
+        }
+        rsp = json.loads(
+            self.raw(
+                "tx",
+                "icaauth",
+                "register-account",
+                connid,
+                "-y",
+                **(default_kwargs | kwargs),
+            )
+        )
+        if rsp["code"] == 0:
+            rsp = self.event_query_tx_for(rsp["txhash"])
+        return rsp
+
     def ica_submit_tx(self, connid, tx, **kwargs):
         default_kwargs = {
             "home": self.data_dir,
@@ -1289,6 +1311,28 @@ class CosmosCLI:
                 "ica",
                 "controller",
                 "send-tx",
+                connid,
+                tx,
+                "-y",
+                **(default_kwargs | kwargs),
+            )
+        )
+        if rsp["code"] == 0:
+            rsp = self.event_query_tx_for(rsp["txhash"])
+        return rsp
+
+    def icaauth_submit_tx(self, connid, tx, **kwargs):
+        default_kwargs = {
+            "home": self.data_dir,
+            "node": self.node_rpc,
+            "chain_id": self.chain_id,
+            "keyring_backend": "test",
+        }
+        rsp = json.loads(
+            self.raw(
+                "tx",
+                "icaauth",
+                "submit-tx",
                 connid,
                 tx,
                 "-y",
