+++ integration_tests/test_ica.py
@@ -1,0 +1,77 @@
+import json
+
+import pytest
+
+from .ibc_utils import (
+    funds_ica,
+    prepare_network,
+    wait_for_check_channel_ready,
+    wait_for_check_tx,
+)
+
+
+@pytest.fixture(scope="module")
+def ibc(request, tmp_path_factory):
+    "prepare-network"
+    name = "ibc"
+    path = tmp_path_factory.mktemp(name)
+    network = prepare_network(path, name, False)
+    yield from network
+
+
+def test_ica(ibc, tmp_path):
+    connid = "connection-0"
+    cli_host = ibc.chainmain.cosmos_cli()
+    cli_controller = ibc.cronos.cosmos_cli()
+
+    print("register ica account")
+    rsp = cli_controller.icaauth_register_account(
+        connid, from_="signer2", gas="400000", fees="100000000basetcro"
+    )
+    assert rsp["code"] == 0, rsp["raw_log"]
+    port_id, channel_id = next(
+        (
+            evt["attributes"][0]["value"],
+            evt["attributes"][1]["value"],
+        )
+        for evt in rsp["events"]
+        if evt["type"] == "channel_open_init"
+    )
+    print("port-id", port_id, "channel-id", channel_id)
+
+    wait_for_check_channel_ready(cli_controller, connid, channel_id)
+
+    print("query ica account")
+    ica_address = cli_controller.ica_query_account(
+        connid, cli_controller.address("signer2")
+    )["interchain_account_address"]
+    print("ica address", ica_address)
+
+    funds_ica(cli_host, ica_address)
+    num_txs = len(cli_host.query_all_txs(ica_address)["txs"])
+
+    # generate a transaction to send to host chain
+    generated_tx = tmp_path / "generated_tx.txt"
+    generated_tx_msg = cli_host.transfer(
+        ica_address, cli_host.address("signer2"), "0.5cro", generate_only=True
+    )
+
+    print(generated_tx_msg)
+    generated_tx.write_text(json.dumps(generated_tx_msg))
+
+    # submit transaction on host chain on behalf of interchain account
+    rsp = cli_controller.icaauth_submit_tx(
+        connid,
+        generated_tx,
+        from_="signer2",
+    )
+    assert rsp["code"] == 0, rsp["raw_log"]
+    packet_seq = next(
+        int(evt["attributes"][4]["value"])
+        for evt in rsp["events"]
+        if evt["type"] == "send_packet"
+    )
+    print("packet sequence", packet_seq)
+    wait_for_check_tx(cli_host, ica_address, num_txs)
+    # check if the funds are reduced in interchain account
+    assert cli_host.balance(ica_address, denom="basecro") == 50000000
